reorganized client side of file structure
