+++ zillowgame/client/public/index.html
@@ -5,18 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
-      name="Zillow Game"
+      name="Z-Guess"
       content="Beat the Zestimate!" />
     <title>Beat The Zestimate!</title>
   </head>
   <body>
     <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.      
-    -->
   </body>
 </html>
