+++ zillowgame/client/src/config/ServerConfig.mjs
@@ -1,6 +1,7 @@
 export const AddressDeploy = '';
 export const PORT = 5005;
 export const AddressLocal = `http://localhost:${PORT}`;
+export const SpotifyAddress = 'NOTIMPLEMENTED'
 export const AddressInUse = AddressLocal;
